+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## Unreleased
 
 - Add optional `compressedSize` and `uncompressedSize` params to `Span.addMessageEvent`
+- Add support for ```tags```, ```status``` and ```annotation``` in Zipkin exporter.
 
 ## 0.0.9 - 2019-02-12
 - Add Metrics API.